All tests now passes !  Code is now ready for server deployment and user testing !

+++ client/mumble-plugin/lib/audio_modern.cpp
@@ -115,7 +115,7 @@ void fgcom_audio_filter(int highpass_cutoff, int lowpass_cutoff, float *outputPC
             }
             
             // Apply lowpass filter if specified
-            if (lowpass_cutoff > 0 && lowpass_cutoff < sampleRateHz / 2) {
+            if (lowpass_cutoff > 0 && lowpass_cutoff < static_cast<int>(sampleRateHz / 2)) {
                 sample = AudioProcessing::lowpassFilter.process(sample);
             }
